change Long to Integer

+++ web/WEB-INF/classes/src/org/oscarehr/PMmodule/web/admin/ProgramManagerAction.java
@@ -285,7 +285,7 @@ public class ProgramManagerAction extends BaseAction {
             return list(mapping, form, request, response);
         }
 
-        int numQueue = programQueueManager.getActiveProgramQueuesByProgramId(Long.valueOf(id)).size();
+        int numQueue = programQueueManager.getActiveProgramQueuesByProgramId(Integer.valueOf(id)).size();
         if (numQueue > 0) {
             ActionMessages messages = new ActionMessages();
             messages.add(ActionMessages.GLOBAL_MESSAGE, new ActionMessage("program.delete.queue", name, String.valueOf(numQueue)));
@@ -649,7 +649,7 @@ public class ProgramManagerAction extends BaseAction {
                     setEditAttributes(request, String.valueOf(program.getId()));
                     return mapping.findForward("edit");
                 }
-                int numQueue = programQueueManager.getActiveProgramQueuesByProgramId((long) program.getId()).size();
+                int numQueue = programQueueManager.getActiveProgramQueuesByProgramId(program.getId()).size();
                 if (numQueue > 0) {
                     ActionMessages messages = new ActionMessages();
                     messages.add(ActionMessages.GLOBAL_MESSAGE, new ActionMessage("program.client_in_the_queue", program.getName(), String.valueOf(numQueue)));
@@ -740,7 +740,7 @@ public class ProgramManagerAction extends BaseAction {
         if (this.isCancelled(request)) {
             return list(mapping, form, request, response);
         }
-        access.setProgramId(program.getId().longValue());
+        access.setProgramId(program.getId());
 
         if (programManager.getProgramAccess(String.valueOf(access.getProgramId()), access.getAccessTypeId()) != null) {
             ActionMessages messages = new ActionMessages();
@@ -816,7 +816,7 @@ public class ProgramManagerAction extends BaseAction {
         if (this.isCancelled(request)) {
             return list(mapping, form, request, response);
         }
-        provider.setProgramId(program.getId().longValue());
+        provider.setProgramId(program.getId());
 
         if (programManager.getProgramProvider(provider.getProviderNo(), String.valueOf(program.getId())) != null) {
             ActionMessages messages = new ActionMessages();
@@ -894,7 +894,7 @@ public class ProgramManagerAction extends BaseAction {
 
             request.setAttribute("admissions", admissionManager.getCurrentAdmissionsByProgramId(programId));
             request.setAttribute("accesses", programManager.getProgramAccesses(programId));
-            request.setAttribute("queue", programQueueManager.getActiveProgramQueuesByProgramId(Long.valueOf(programId)));
+            request.setAttribute("queue", programQueueManager.getActiveProgramQueuesByProgramId(Integer.valueOf(programId)));
             request.setAttribute("programFirstSignature",programManager.getProgramFirstSignature(Integer.valueOf(programId)));
         }
